refactor: put validation for revoke and approveProcessedData into its own class and rename SequenceValidatorService
